% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3_exportISS.R
\name{exportISS}
\alias{exportISS}
\title{Export ISS data in different formated file.}
\usage{
exportISS(file, toexport = "WB")
}
\arguments{
\item{file}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}. See details.}

\item{toexport}{To which formate data will export. See details.}
}
\description{
Export ISS data to different formated file which are compitible for other software or R package.
}
\details{
At present only non-segmentated file (file class \link[MolDia]{MolDiaISS_nonsegment})is supported for 
         exporting file formate compitible with R package wholebrainsoftware.
         
         toexport only support "WB" argument at this moment, which is compitible with  wholebrainsoftware.
}
\examples{
###### Reading non-segmentated file
mdata <- readISS(file = system.file("extdata", "nonSeg_QT_0.35_0_details.csv", package="MolDia"), segment = FALSE,
                 centX = "PosX", centY = "PosY", rmNAgene = "NNNN", gene= c("Gdf7","WNT1","Pak3","Tfap2a"))
mywb  <- exportISS(file = mdata, toexport= "WB")
}
