% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0_RCA_readData_summary.R
\name{genesSummary}
\alias{genesSummary}
\title{Summary of specific gene of interest of RCA data}
\usage{
genesSummary(data, gene = NULL)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}}

\item{gene}{gene of interest}
}
\value{
Barplot
}
\description{
Summary of specific gene of interest of RCA data
}
\examples{
data_1 <- readRCA(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX ="centroidX", centY="centroidY" )
res <- genesSummary(data = data_1, gene = "Actb.L")

}
