% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.5_ISS_clustCompare.R
\name{ISS_clustCompare}
\alias{ISS_clustCompare}
\title{Compare among ISS clusters based on Random Forest algorithm}
\usage{
ISS_clustCompare(preditorData, prediction, ntree = 100, method = "prob")
}
\arguments{
\item{preditorData}{Data to be used to compare. Input data in class MolDiaISS. Output of \link[MolDia]{readISS}
and must be clustered.}

\item{prediction}{Data that need to compare. Input data in class MolDiaISS. Output of \link[MolDia]{readISS}
and must be clustered.}

\item{ntree}{Number of trees to grow. \link[randomForest]{randomForest}. See This should not be set to
too small a number, to ensure that every input row gets predicted at least a few times.}

\item{method}{Methods to return comarison results. Possible values are "vote" and "prob". Defaut is "prob".}
}
\description{
Compare two cluster and find their similarity (Supervised).
}
\details{
This function will be apply to compare Two cluster by Random forest algorithm.
}
\note{
This function need optimization to run properly. Work fine with small data but not with big data .
}
\examples{
## Read data: Left and right HC
left_hypo  <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
right_hypo <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"),

                                     cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
## Data preprocessing
left_hypo  <- ISS_preprocess(data = left_hypo, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)
right_hypo <- ISS_preprocess(data = right_hypo, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
left_hypo   <- ISS_cluster(data =  left_hypo, pc = 0.9, resolution = 0.05)
right_hypo  <- ISS_cluster(data = right_hypo, pc = 0.9, resolution = 0.05)

## Cluster compare
comapre_left_right <- ISS_clustCompare(preditorData = left_hypo, prediction = right_hypo, ntree = 10)
comapre_right_left <- ISS_clustCompare(preditorData = right_hypo, prediction = left_hypo, ntree = 10)


}
