% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.1_ISS_cluster.R
\name{ISS_cluster}
\alias{ISS_cluster}
\title{Cluster ISS data based on different methods.}
\usage{
ISS_cluster(data, method = "seurat", DEGmethod = "seurat", pc = 1,
  cluster_id = NULL, resolution = 0.3, algorithm = 1,
  do.norm = TRUE, do.scale = TRUE, k.param = 30)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}.}

\item{method}{Method of clustering. SEURAT, backspin, monocle.}

\item{DEGmethod}{Methods to find DE genes.NULL means use all genes.}

\item{pc}{Desired percent of variance (0 to 1) to be explained by PCA. Default is 1 (All PC will use).}

\item{cluster_id}{Re-cluster clustreded data. Numeric input. Default is NULL.}

\item{resolution}{Value of the resolution parameter, use a value above (below)
1.0 if you want to obtain a larger (smaller) number of communities.
Default is 0.3.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm;
2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm). Default is 1.}

\item{do.norm}{Do normalization}

\item{do.scale}{Do scalling}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}
}
\description{
Cluster ISS data by different methods used in SEURAT, MONOCLE, BackSPIN.
}
\examples{
## Reading data
left_hypo <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId", centX = "centroid_x", centY = "centroid_y")

## Arrange marker gene
data(marker_gene)
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro)

## Barplot of Neuronal marker gene and extract those cells only
neuron_group <- ISS_barplot(data = left_hypo, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)
## Data preprocessing
neuron_group <- ISS_preprocess(data = neuron_group, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
neuron_group_clust  <- ISS_cluster (data = neuron_group, method = "seurat",
                                    pc = 0.9, resolution = 0.3)

## Re-cluster specific cluster
re_clust  <- ISS_cluster (data = neuron_group_clust, method = "seurat",
                                 pc = 0.9, cluster_id = 3, resolution = 1)

}
