% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3_ISS_ClustSelect.R
\name{ISS_ClustSelect}
\alias{ISS_ClustSelect}
\title{Select cluster of interest after clustering.}
\usage{
ISS_ClustSelect(data, cluster_id = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}.}

\item{cluster_id}{Cluster to select}
}
\description{
Select cluster of interest after clustering
}
