% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.1_ISS_compare.R
\name{ISS_compare}
\alias{ISS_compare}
\title{Compare multiple ISS data based on gene}
\usage{
ISS_compare(..., logdata = FALSE, label = TRUE, levelCI = 0.99,
  live = FALSE)
}
\arguments{
\item{...}{Input data in class MolDiaISS. At least 2 dataset or more. Output of \link[MolDia]{readISS}.}

\item{logdata}{log2 apply to data or not. Deafult is FALSE.}

\item{label}{Label each point. Default is TRUE.}

\item{levelCI}{Level of confidence interval to use. Default is 0.95}

\item{live}{Show plot in interactive mode. Default is FALSE}
}
\description{
Compare multiple ISS data based on number of reads per gene and observe their goodness of fit and correlation.
}
\examples{
hcleft   <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcright  <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"), cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcleft1  <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcright1 <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"), cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
res <- ISS_compare(hcleft, hcright,hcleft1,hcright1, label = F, levelCI = 0.99, live = F, logdata = FALSE)
res <- ISS_compare(hcleft, hcright, label = T, levelCI = 0.8, live = F, logdata = TRUE)

}
