% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1_ISS_deg.R
\name{ISS_deg}
\alias{ISS_deg}
\title{Find differentially express (DE) gene by different methods}
\usage{
ISS_deg(data, DEGmethod = "seurat")
}
\arguments{
\item{data}{Input data matrix. Column is gene and row is cells.}

\item{DEGmethod}{Methods to find DE genes.}
}
\description{
Find differentially express (DE) gene by different methods
}
\note{
This function is in-complete and needs further development by adding 
      different methods of finding DE optimized for ISS data.
}
\examples{
## Reading data
ex_data <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                   cellid = "CellId", centX = "centroid_x", centY = "centroid_y")

res <- ISS_deg (data = ex_data)

}
