% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.2_ISS_filter.R
\name{ISS_filter}
\alias{ISS_filter}
\title{Filter ISS data based on poisson distibution}
\usage{
ISS_filter(data, data_mean = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}}

\item{data_mean}{Expected mean of number of reads press cell. Default is NULL. See details.}
}
\value{
Number of reads with peobability
}
\description{
This function estimate the probabity of number of reads per cell in a specific range
             with desired mean (Rate or number or reads per cell) by Poisson distibution. By default
             current fuction calculate the mean from data.
}
\examples{
data_1 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", , centX = "centroidX", centY = "centroidY")
res    <- ISS_filter(data = data_1, data_mean = 10)

}
