% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.3_ISS_genesSummary.R
\name{ISS_genesSummary}
\alias{ISS_genesSummary}
\title{Summary of specific gene of interest in ISS data}
\usage{
ISS_genesSummary(data, gene = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}}

\item{gene}{gene of interest}
}
\value{
Barplot
}
\description{
Summary of specific gene of interest of RCA data
}
\examples{
data_1 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX ="centroidX", centY="centroidY" )
res <- ISS_genesSummary(data = data_1, gene = "Actb.L")

}
