% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1_ISS_pca.R
\name{ISS_pca}
\alias{ISS_pca}
\title{Principle component analyis on ISS data}
\usage{
ISS_pca(data, pc = 1, DEGmethod = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}.}

\item{pc}{Desired percent of variance to be explained by PCA. Default is 1 which means 100 percent variation explained.}

\item{DEGmethod}{Methods for finding differentially expressed (DE) genes.}
}
\description{
Principle component analysis on any data in class MolDiaISS.
}
\examples{
## Reading data
left_hypo <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                     cellid = "CellId", centX = "centroid_x", centY = "centroid_y")
                  
## Arrange marker gene
data(marker_gene)
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                                      nonneuron = marker_gene$genr_nonneuro)

## Barplot of Neuronal marker gene and extract those cells only
neuron_group <- ISS_barplot(data = left_hypo, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)

## Preprocess data
neuron_group <- ISS_preprocess(data = neuron_group, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)
                               
## Apply principle component analysis
res <- ISS_pca(data = neuron_group, pc = 0.9)

}
