% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.0_RCA_pieplot.R
\name{ISS_pieplot}
\alias{ISS_pieplot}
\title{Venn-pie chart on RCA data based on genes of interest}
\usage{
ISS_pieplot(data, gene = NULL, with_gene = NULL, without_gene = NULL)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{gene}{Gene of interest. Default is NULL.}
}
\value{
The Venn-pie plot.

        Result in class \link[MolDia]{readRCA}.
}
\description{
Venn-pie chart on RCA data based on genes of interest
}
\examples{
########## Reading data
data_1      <- readRCA(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                       cellid = "CellID",centX = "centroidX", centY = "centroidY")

## Define marker gene group
marker_gene <- data_1@gene
c_rna <- marker_gene[grepl(".C",marker_gene)] # Circular RNA
l_rna <- setdiff(marker_gene,c_rna)           # Linear RNA

## Neuronal marker
Pyramidal   <- c("Nrn1.L","Pcp4.L")
Interneuron <- c("Sst.L","Pvalb.L","Ndnf.L","Vip.L","Sncg.L","Cck.L")
Other       <- c("Map2.L")
Neuron      <- c(Pyramidal,Interneuron,Other)

## Non-Neuronal marker
Oligodendrocyte <- c("Plp1.L","Enpp2.L")
Astrocytes  <- c("Gfap.L","Mfge8.L","Aldh1l1.L","S100b.L")
Nonneuron   <- c(Oligodendrocyte,Astrocytes)
marker_gene1 <- list(Neuron = Neuron,
                     Nonneuron = Nonneuron,
                     Cir_rna = c_rna,
                     Lin_rna = c(setdiff(setdiff(l_rna,Neuron),Nonneuron)))

####  Select cell with at least 3 neuronal genes
mydata <- RCA_barplot(data = data_1, gene = marker_gene1, gene.target = 1,
                      at.least.gene = 3, show.same.gene = FALSE)

#####  Pie chart of all Circular mRNA in cell with at least 3 neuronal genes.
genes <- marker_gene1[[1]]
res   <- ISS_pieplot (data = mydata, gene = genes)

}
\author{
Mohammad Tanvir Ahamed
}
