% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1_ISS_preprocess.R
\name{ISS_preprocess}
\alias{ISS_preprocess}
\title{Preprocess ISS data based on different normalization method.}
\usage{
ISS_preprocess(data, normalization.method = NULL, scale.factor = 10000,
  do.scale = FALSE, do.center = FALSE, display.progress = FALSE)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}}

\item{normalization.method}{Method for normalization. Default is NULL.
Available methods are "LogNormalize", "RankNormalize" and "QuantileNormalize".}

\item{scale.factor}{Sets the scale factor for cell-level normalization.}

\item{do.scale}{Whether to scale the data. See details.}

\item{do.center}{Whether to center the data. See details.}

\item{display.progress}{Displays a progress bar}
}
\value{
Return a object in MolDiaISS with value in slot norm.data or scale.data
}
\description{
Pre-process RCA data interms of normalization, scalling and centering
}
\details{
Setting center to TRUE will center the expression for each gene by subtracting the average expression for that gene.
         Setting scale to TRUE will scale the expression level for each gene by dividing the centered gene expression levels
         by their standard deviations if center is TRUE and by their root mean square otherwise.
}
\examples{
mydata  <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                   cellid = "CellID", centX = "centroidX", centY = "centroidY")
res     <- ISS_preprocess(mydata, normalization.method = "QuantileNormalize", do.scale = FALSE, do.center = FALSE)




}
\references{
Van den Berg, R. A., Hoefsloot, H. C., Westerhuis, J. A., Smilde, A. K., & van der Werf, M. J. (2006). Centering, scaling, and
transformations: improving the biological information content of metabolomics data. BMC Genomics, 7, 142
http://doi.org/10.1186/1471-2164-7-142
}
