% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISS_ratiocor.R
\name{ISS_ratiocor}
\alias{ISS_ratiocor}
\title{Calculate and plot correlation and ratio of total reads between genes}
\usage{
ISS_ratiocor(data, gene = marker_gene, select_gene = NULL,
  errorbar = TRUE, plty = "both", logratio = TRUE, sig.level = 0.05,
  main = "", stat = "sum")
}
\arguments{
\item{data}{Data in list of groups with group name.   Individual data is in class RCA_class. Output of \link[MolDia]{readRCA}}

\item{gene}{Object in vector or list formate. In list formated input every list element is a group of
interested genes.}

\item{select_gene}{Select gene of interest, Default is NULL i.e. all genes in gele list.}

\item{errorbar}{Show error bar or not. Default is TRUE}

\item{plty}{Show plot type. Available balue is "corr", "ratio" and "both". Default is "both".}

\item{logratio}{Taking log2 of the ratio of gene's total reads count.}

\item{sig.level}{confidence level for the returned confidence interval. Currently only used for 
the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}

\item{main}{Title of the plot}

\item{stat}{Mode of operation. Possible value is "sum", "gene" and "present". Default is "sum".}
}
\description{
Calculate and plot correlation and ratio of total reads between genes
}
\examples{
## Read data: Left and right HC
hc_left  <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hc_right <- readRCA(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")

## Arrange marker gene
data(marker_gene)
marker_gene <- marker_gene
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro) 

res <- ISS_ratiocor(data = list(Left_HC = c(hc_left),Right_HC = c(hc_right)), gene = mark_gene, plty = "ratio")
}
