% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.4_ISS_readsSummary.R
\name{ISS_readsSummary}
\alias{ISS_readsSummary}
\title{Summary of ISS data and expression}
\usage{
ISS_readsSummary(data, readlimit = 10, text.size = 6, intervel.dep = NULL,
  coexp = FALSE)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}}

\item{readlimit}{Total number of reads per cell to consider}

\item{text.size}{Text size of Gene name.}

\item{intervel.dep}{Vector of values between 0 and 100 , where you want to set interval. If one want to define the range like
0 to 5, 5 to 20, 20 to 40, 40 to 60, 60 to 100 then the input valuse will be c(5,20,40,60)}

\item{coexp}{Same gene coexpression. Default is FALSE. TRUE means at least 2 reads.}
}
\value{
Summary of cells after reads delete.
}
\description{
Summary of ISS data and expression
}
\examples{
data_1 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX = "centroidX", cent = "centroidY")
res    <- ISS_readsSummary(data = data_1, readlimit = 10, text.size = 6, intervel.dep = NULL, coexp = TRUE)

}
\author{
Mohammad Tanvir Ahamed
}
