% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.2_ISS_ratiocor.R
\name{ISS_sumstat}
\alias{ISS_sumstat}
\title{Calculate summary statstics of data}
\usage{
ISS_sumstat(data, gene, stat = "sum")
}
\arguments{
\item{data}{Data in dataframe class.}

\item{gene}{Object in vector or list formate. In list formated input every list element is a group of
interested genes.}

\item{stat}{Mode of operation. Possible value is "sum", "gene" and "present". Default is "sum".}
}
\value{
Output will be in dataframe class with a extra column names "total_reads" wish is the total numbe rof reads in each cells.
}
\description{
Calculate summary statstics (Total reads, Total number of gene present and Gene present in group) of data
}
\keyword{internal}
