% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_data.R
\docType{data}
\name{marker_gene}
\alias{marker_gene}
\title{Gene set for neuronal marker}
\format{A datset of neuronal gene listed in different category (List).}
\usage{
data(marker_gene)
}
\description{
This is a sample example for neuronal gene marker in different category.
}
\keyword{datasets}
