% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1_MolDiaISS_class.R
\docType{data}
\name{MolDiaISS}
\alias{MolDiaISS}
\title{The RCA class
The RCA class is a main object for all RCA based analysis}
\format{An object of class \code{MolDiaISS} of length 1.}
\usage{
MolDiaISS
}
\description{
A RCA object has the following slot
}
\section{Slots}{

\describe{
\item{\code{data}}{Main data for analysis}

\item{\code{norm.data}}{Normalized data (log-scale).}

\item{\code{scale.data}}{Scaled (default is z-scoring each gene) expression matrix; used for dimmensional reduction and heatmap visualization.}

\item{\code{location}}{Insitu location of RCA cell location on X and Y co-ordinate.}

\item{\code{gene}}{Name of genes in the data.}

\item{\code{cluster}}{Cluster identification of all cells}

\item{\code{cluster.marker}}{Marker gene of each cluster.}

\item{\code{tsne.data}}{tSNE co-ordinate for data.}
}}

\section{Dataclass}{
 Data class section
}

\keyword{datasets}
