% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1_MolDiaISS_class.R
\docType{data}
\name{MolDiaISS_nonsegment}
\alias{MolDiaISS_nonsegment}
\title{The ISS data class for non-segmentated file}
\format{An object of class \code{MolDiaISS_nonsegment} of length 1.}
\usage{
MolDiaISS_nonsegment
}
\description{
The MolDiaISS_nonsegment class is the non segmentated data class.
}
\section{Slots}{

\describe{
\item{\code{reads}}{Reads name of the data.}

\item{\code{gene}}{Name of genes in the data.}

\item{\code{tile}}{ParentCell and tile of the data.}

\item{\code{quality}}{Quality information of the data.}

\item{\code{location}}{In-situ location of the gene in the data}
}}

\keyword{datasets}
