% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCA_clustcompare.R
\name{RCA_clustcompare}
\alias{RCA_clustcompare}
\title{Compare among Single cell cluster based Random forest algorithm}
\usage{
RCA_clustcompare(preditorData, prediction, ntree = 500, method = "prob")
}
\arguments{
\item{preditorData}{Data to be used to compare. Input data in class MolDiaISS. Output of \link[MolDia]{readISS}
and must be clustered.}

\item{prediction}{Data that need to compare. Input data in class MolDiaISS. Output of \link[MolDia]{readISS}
and must be clustered.}

\item{ntree}{Number of trees to grow. \link[randomForest]{randomForest}. See This should not be set to
too small a number, to ensure that every input row gets predicted at least a few times.}

\item{method}{Methods to return comarison results. Possible valus is "vote" and "prob". Defaut is "prob".}
}
\description{
Compare two two cluster and find their similarity (Supervised).
}
\details{
This function will be apply to compare Two cluster by Random forest algorithm.
}
\examples{
## Read data: Left and right HC
hc_left  <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hc_right <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")

## Arrange marker gene
data(marker_gene)
marker_gene <- marker_gene
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro)
## Barplot of Neuronal marker gene and extract those cells only
hc_left  <- RCA_barplot(data = hc_left, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)
hc_right <- RCA_barplot(data = hc_right, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)

## Data preprocessing
hc_left  <- RCA_preprocess(data = hc_left, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)
hc_right <- RCA_preprocess(data = hc_right, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
hc_left   <- RCA_cluster(data =  hc_left, pc = 0.9, resolution = 0.3)
hc_right  <- RCA_cluster(data = hc_right, pc = 0.9, resolution = 0.3)

## Get cluster marker
hc_left  <- RCA_marker(data = hc_left, topgene =5, test.use="bimod",
                      main = "hc_left")
hc_right <- RCA_marker(data = hc_right, topgene =5, test.use="bimod",
                      main = "hc_left")

## Cluster compare
my_comapre_1 <- RCA_clustcompare(preditorData = hc_left, prediction = hc_right)
my_comapre_2 <- RCA_clustcompare(preditorData = hc_right, prediction = hc_left)


}
