% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18.0_RCA_de.R
\name{RCA_DE}
\alias{RCA_DE}
\title{Find differentially express gene by different methods}
\usage{
RCA_DE(data, DEGmethod = "seurat")
}
\arguments{
\item{data}{Input data matrix. Column is gene and row is cells.}

\item{DEGmethod}{Methods to find DE genes.}
}
\description{
Find differentially express gene by different methods
}
