% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_RCA_plot.R
\name{RCA_map}
\alias{RCA_map}
\title{Plot function for RCA_class data.}
\usage{
RCA_map(data, what = "cell", xlab = "centroid_x", ylab = "centroid_y",
  main = "Main plot", image = TRUE, live = FALSE, label.topgene = NULL,
  gene = NULL)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{what}{What to plot. Values can be "cell", "cluster", "tsne". Default is "cell".}

\item{xlab}{Label of x-axis}

\item{ylab}{Label of y-axis}

\item{main}{Main title}

\item{image}{Plot image. Default is TRUE.}

\item{live}{Plot interactive image. Default is FALSE.}

\item{label.topgene}{Active only when "what = cluster or tsne". Number of genes to label each cluster. Only work when data is clustered and clusted marker
has identified in cluster.marker slot of input data.}

\item{gene}{Gene of interest, Now limited to max 6 genes.}
}
\description{
Map RCA data based on cell, cluster or tSNE,
}
