% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20.0_RCA_ClusterMarker.R
\name{RCA_marker}
\alias{RCA_marker}
\title{Find Cluster marker of in-situ RCA data.}
\usage{
RCA_marker(data, topgene = 5, test.use = "bimod", marker.sig = 0.005,
  only.pos = TRUE, main = "")
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{topgene}{Desired number of top gene ineach cluster to show in summary result.}

\item{test.use}{Denotes which test to use. Seurat currently implements "bimod"
(likelihood-ratio test for single cell gene expression, McDavid et al., Bioinformatics, 2013, default), "roc"
(standard AUC classifier), "t" (Students t-test), and "tobit" (Tobit-test for differential gene expression,
as in Trapnell et al., Nature Biotech, 2014). For details see \link[Seurat]{FindAllMarkers}}

\item{marker.sig}{Lower value will identify less significant marker. Default is 0.005}

\item{only.pos}{Only return positive markers (TRUE by default)}

\item{main}{Title of the plot.}
}
\value{
A list of cluster with putative ranked markers and associated statistics in slot cluster.marker of main data.
        Also clusterwise figure in barplot and heatmap of desired top genes.
}
\description{
Find Cluster marker of in-situ RCA data and plot significant gene cluster wise by barplot and heatmap.
}
\examples{
## Reading data
data_3 <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId", centX = "centroid_x", centY = "centroid_y")

## Arrange marker gene
data(marker_gene)
marker_gene <- marker_gene
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro)

## Barplot of Neuronal marker gene and extract those cells only
neuron_group <- RCA_barplot(data = data_3, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)
## Data preprocessing
neuron_group <- RCA_preprocess(data = neuron_group, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
neuron_group_clust  <- RCA_cluster (data = neuron_group, method = "seurat",
                                    pc = 0.9, resolution = 0.4)
## Get cluster marker
neuron_group_clust_marker <- RCA_marker(data = neuron_group_clust, topgene =15,
                                        test.use="bimod", main = "")

}
