% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17.0_RCA_preprocess.R
\name{RCA_preprocess}
\alias{RCA_preprocess}
\title{Pre-process RCA data.}
\usage{
RCA_preprocess(data, normalization.method = "LogNormalize",
  scale.factor = 10000, do.scale = TRUE, do.center = TRUE,
  display.progress = FALSE)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}}

\item{normalization.method}{Method for normalization. Default is log-normalization (LogNormalize).
Use NULL if one dont apply normalization. More methods to be added very shortly.}

\item{scale.factor}{Sets the scale factor for cell-level normalization.}

\item{do.scale}{Whether to scale the data. See details.}

\item{do.center}{Whether to center the data. See details.}

\item{display.progress}{Displays a progress bar}
}
\value{
Return a object in RCA_class with value in slot norm.data or scale.data
}
\description{
Pre-process RCA data interms of normalization, scalling and centering
}
\details{
Setting center to TRUE will center the expression for each gene by subtracting the average expression for that gene.
         Setting scale to TRUE will scale the expression level for each gene by dividing the centered gene expression levels
         by their standard deviations if center is TRUE and by their root mean square otherwise.
}
\examples{
data_1 <- readRCA(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID")
scale_data_1 <- RCA_preprocess(data_1)

}
\references{
Van den Berg, R. A., Hoefsloot, H. C., Westerhuis, J. A., Smilde, A. K., & van der Werf, M. J. (2006). Centering, scaling, and
transformations: improving the biological information content of metabolomics data. BMC Genomics, 7, 142.
http://doi.org/10.1186/1471-2164-7-142
}
