% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19.1.1_RCA_Cluster_SEURAT.R
\name{RCA_seruat_cluster}
\alias{RCA_seruat_cluster}
\title{"RCA_seruat_cluster"
Cluster in-situ RCA data by SEURAT.}
\usage{
RCA_seruat_cluster(data, pc = NULL, cluster_id = NULL, resolution = 0.3,
  algorithm = 1, DEGmethod = "seurat", k.param = 30)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{pc}{Desired percent of variance (0 to 1) to be explained by PCA. Default in NULL (All PC will use).}

\item{cluster_id}{Re-cluster clustreded data. Numeric input. Default is NULL.}

\item{resolution}{Value of the resolution parameter, use a value above (below)
1.0 if you want to obtain a larger (smaller) number of communities.
Default is 0.3.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm;
2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm). Default is 1.}

\item{DEGmethod}{Methods to find DE genes.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm}
}
\description{
"RCA_seruat_cluster"
Cluster in-situ RCA data by SEURAT.
}
\examples{
## Reading data
data_3 <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId", centX = "centroid_x", centY = "centroid_y")

## Arrange marker gene
data(marker_gene)
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro)

## Barplot of Neuronal marker gene and extract those cells only
neuron_group <- RCA_barplot(data = data_3, gene = mark_gene, gene.target = 2,
                            at.least.gene = 8, gene.show = 2)
## Data preprocessing
neuron_group <- RCA_preprocess(data = neuron_group, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
neuron_group_clust  <- RCA_seruat_cluster (data = neuron_group, pc = 0.9, resolution = 0.4)

## Re-cluster specific cluster
# re_clust  <- RCA_seruat_cluster (data = neuron_group_clust, pc = 0.9,
#                                  cluster_id = 0, resolution = 0.5)

}
\keyword{internal}
