% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19.1.2_RCA_Cluster_Spatial.R
\name{RCA_spatial}
\alias{RCA_spatial}
\title{Spatial clustering for ISS.}
\usage{
RCA_spatial(data, neighbor.points = 9)
}
\description{
ISS spatial clustering
}
\examples{
## Reading data
ex_data <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                   cellid = "CellId", centX = "centroid_x", centY = "centroid_y")
myres <- RCA_spatial(ex_data, neighbor.points = 4)

}
