% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1_readISS.R
\name{readISS}
\alias{readISS}
\title{Read RCA data}
\usage{
readISS(file, cellid = "CellID", centX = NULL, centY = NULL,
  genepos = NULL, geneposOPT = "OR", rpc = 1, rpg = 1, gene = NULL,
  nogene = NULL)
}
\arguments{
\item{file}{File name in CSV formate, data in "data.frame" class and data in "MolDiaISS" class (Output of \link[MolDia]{readISS}).}

\item{cellid}{String to naming cell. Default is "CellID".}

\item{centX}{Name of X co-ordinate in file. Default is "centroidX"}

\item{centY}{Name of Y co-ordinate in file  Default is "centroidY"}

\item{genepos}{Name of genes to consider for gene positive cells. Default is NULL}

\item{geneposOPT}{Only work when 'genepos' has a value. "AND", "OR" and "NONE" condition for genepos. Default is "OR".}

\item{rpc}{Total reads per cell to be consider. Default is 1.}

\item{rpg}{Total reads per gene to be consider. Default is 1.}

\item{gene}{Gene name to include in data. Default is NULL.}

\item{nogene}{Gene name to exclude from data. Default is NULL.}
}
\value{
Output will be a object in class MolDiaISS. See detail \link[MolDia]{MolDiaISS}
}
\description{
Read RCA data
}
\examples{
##### Reading ISS data in CSV formate
data_1 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX = "centroidX", centY = "centroidY")
data_2 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.40.csv", package="MolDia"),
                  cellid = "CellID",centX = "centroidX", centY = "centroidY")
data_3 <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
data_4 <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")

## Not RUN
## Read RCA data in dataframe formate
#gene <- data_3@gene[1:3]
#data(single_cell)
#single_cell$CellID <- rownames(single_cell)
#Note: Big data. Take long time to load without selected gene
#data_sc <- readISS(file = single_cell, cellid = "CellID",gene= gene) 


}
\author{
Mohammad Tanvir Ahamed
}
