% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0_RCA_readData_summary.R
\name{readRCA}
\alias{readRCA}
\title{Read RCA data}
\usage{
readRCA(file, cellid = "CellID", centX = NULL, centY = NULL,
  genepos = NULL, geneposOPT = "OR", rpc = 1, rpg = 1, gene = NULL,
  nogene = NULL)
}
\arguments{
\item{file}{File name in CSV formate or data in data.frame class}

\item{cellid}{String to naming cell. Default is "CellID".}

\item{centX}{Name of X co-ordinate in file. Default is "centroidX"}

\item{centY}{Name of Y co-ordinate in file  Default is "centroidY"}

\item{genepos}{Name of genes to consider for gene positive cells. Default is NULL}

\item{geneposOPT}{"AND" and "OR" condition for genepos. Default is "OR".}

\item{rpc}{Total reads per cell to be consider. Default is 1.}

\item{rpg}{Total reads per gene to be consider. Default is 1.}

\item{gene}{Gene name to include in data. Default is NULL.}

\item{nogene}{Gene name to exclude from data. Default is NULL.}
}
\value{
Output will be a object in class RCA_class. See detail \link[MolDia]{RCA_class}
}
\description{
Read RCA data
}
\examples{
##### Reading RCA data in CSV formate
data_1 <- readRCA(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX = "centroidX", centY = "centroidY")
data_2 <- readRCA(file = system.file("extdata", "CellBlobs_QT_0.40.csv", package="MolDia"),
                  cellid = "CellID")
data_3 <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
data_4 <- readRCA(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"),
                  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")

## Read RCA data in dataframe formate
gene <- data_3@gene

#Note: Big data. Take long time to load without selected gene
## Not RUN
#data(single_cell)
#single_cell$CellID <- rownames(single_cell)
#data_sc <- readRCA(single_cell, gene= gene)


}
\author{
Mohammad Tanvir Ahamed
}
