% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.3_Ex_data.R
\docType{data}
\name{single_cell}
\alias{single_cell}
\title{Single cell RNA seq data}
\format{A datset of single cell RNA seq data.}
\usage{
data(single_cell)
}
\description{
Single cell RNA seq data for mouse hippocampus
}
\keyword{datasets}
