% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2_writeISS.R
\name{writeISS}
\alias{writeISS}
\title{Write ISS data in CSV formate}
\usage{
writeISS(file, output = "", expr = TRUE, location = TRUE)
}
\arguments{
\item{file}{File name in "MolDiaISS" class (Output of \link[MolDia]{readISS}).}

\item{output}{A character string naming a file.}

\item{expr}{Write expression data. TRUE or FALSE needed. Default is TRUE.}

\item{location}{Write location data. TRUE or FALSE needed. Default is TRUE.}
}
\value{
Output will be a CSV formated file.
}
\description{
Write ISS data in CSV formate.
}
\examples{

##### Reading ISS data from CSV formate
data_1 <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                  cellid = "CellID", centX = "centroidX", centY = "centroidY")
writeISS(file = data_1, output  = "test.csv", expr = F, location = T)
}
\author{
Mohammad Tanvir Ahamed
}
