% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1.1_ISS_compare.R
\name{ISS_compare}
\alias{ISS_compare}
\title{Compare ISS data}
\usage{
ISS_compare(..., logdata = FALSE, label = TRUE, levelCI = 0.95,
  live = TRUE)
}
\description{
Compare multiple ISS data
}
\examples{
hc_left  <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hc_right <- readRCA(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"), cellid = "CellId",centX = "centroid_x", centY = "centroid_y")

kk<- ISS_compare(hc_left,hc_right,hc_left,hc_right, label = F, levelCI = 0.99, live = F, logdata = FALSE)

}
