% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISS_compare.R
\name{ISS_compare}
\alias{ISS_compare}
\title{Compare ISS data}
\usage{
ISS_compare(..., logdata = FALSE, label = TRUE, levelCI = 0.95,
  live = FALSE)
}
\arguments{
\item{...}{Input data in class RCA_class. At least 2 dataset or more. Output of \link[MolDia]{readISS}.}

\item{logdata}{log2 apply to data or not. Deafult is FALSE.}

\item{label}{Label each point. Default is TRUE.}

\item{levelCI}{Level of confidence interval to use. Default is 0.95}

\item{live}{Show plot in interactive mode. Default is FALSE}
}
\description{
Compare multiple ISS data
}
\examples{
hcleft   <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcright  <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"), cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcleft1  <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),  cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
hcright1 <- readISS(file = system.file("extdata", "Hypocampus_right.csv", package="MolDia"), cellid = "CellId",centX = "centroid_x", centY = "centroid_y")
kk<- ISS_compare(hcleft, hcright, hcleft1,hcright1,label = F, levelCI = 0.99, live = F, logdata = FALSE)

}
