% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.5_ISS_pieplot.R
\name{ISS_pieplot}
\alias{ISS_pieplot}
\title{Venn-pie chart on ISS data based on genes of interest}
\usage{
ISS_pieplot(data, gene = NULL, with_gene = NULL, without_gene = NULL,
  segmentcol = "EEEEEE", randomseed = 10, colorpalette = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}.}

\item{gene}{Gene of interest. It could be a vector of gene or a list of genes groups. Default is NULL means takes all genes
in the given dataset.}

\item{with_gene}{From gene of interest , select only cells only with these genes.}

\item{without_gene}{From gene of interest , select only cells only without these genes.}

\item{segmentcol}{Hex color code of the overlap segment.}

\item{randomseed}{Ramdom seed to change color schema.}

\item{colorpalette}{Default is NULL. A vector of value . Can be applicabile with selected element with gene list.}
}
\value{
The Venn-pie plot.

        Number of cells having specific gene profile.
}
\description{
Venn-pie chart on ISS data based on genes of interest
}
\note{
Maximum 256 genes (256 color) can be ploted with this version at this moment.
}
\examples{
########## Reading data
data_1      <- readISS(file = system.file("extdata", "CellBlobs_QT_0.35.csv", package="MolDia"),
                       cellid = "CellID",centX = "centroidX", centY = "centroidY")

## Define marker gene group
marker_gene <- data_1@gene
c_rna <- marker_gene[grepl(".C",marker_gene)] # Circular RNA
l_rna <- setdiff(marker_gene,c_rna)           # Linear RNA

## Neuronal marker
Pyramidal   <- c("Nrn1.L","Pcp4.L")
Interneuron <- c("Sst.L","Pvalb.L","Ndnf.L","Vip.L","Sncg.L","Cck.L")
Other       <- c("Map2.L")
Neuron      <- c(Pyramidal,Interneuron,Other)

## Non-Neuronal marker
Oligodendrocyte <- c("Plp1.L","Enpp2.L")
Astrocytes  <- c("Gfap.L","Mfge8.L","Aldh1l1.L","S100b.L")
Nonneuron   <- c(Oligodendrocyte,Astrocytes)
marker_gene1 <- list(Neuron = Neuron,
                     Nonneuron = Nonneuron,
                     Cir_rna = c_rna,
                     Lin_rna = c(setdiff(setdiff(l_rna,Neuron),Nonneuron)))

####  Select cell with at least 3 neuronal genes
mydata <- RCA_barplot(data = data_1, gene = marker_gene1, gene.target = 1,
                      at.least.gene = 3, show.same.gene = FALSE)

#####  Pie chart of all Circular mRNA in cell with at least 3 neuronal genes.
genes <- marker_gene1[[1]]
res   <- ISS_pieplot (data = mydata, gene = genes, without_gene = "QKI.L", segmentcol = "EEEEEE")

}
\author{
Mohammad Tanvir Ahamed
}
