% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1.2_ISS_cor_ratio.R
\name{ISS_ratiocor}
\alias{ISS_ratiocor}
\title{Calculate and plot correlation and ratio of total reads between genes}
\usage{
ISS_ratiocor(data, gene = marker_gene, select_gene = NULL,
  errorbar = TRUE, plty = "both", logratio = TRUE, sig.level = 0.05,
  main = "", stat = "sum")
}
\arguments{
\item{data}{Data in list of groups with group name.   Individual data is in class RCA_class. Output of \link[MolDia]{readRCA}}

\item{gene}{Object in vector or list formate. In list formated input every list element is a group of
interested genes.}

\item{select_gene}{Select gene of interest, Default is NULL i.e. all genes in gele list.}

\item{errorbar}{Show error bar or not. Default is TRUE}

\item{plty}{Show plot type. Available balue is "corr", "ratio" and "both". Default is "both".}

\item{logratio}{Taking log2 of the ratio of gene's total reads count.}

\item{sig.level}{confidence level for the returned confidence interval. Currently only used for 
the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}

\item{main}{Title of the plot}

\item{stat}{Mode of operation. Possible value is "sum", "gene" and "present". Default is "sum".}
}
\description{
Calculate and plot correlation and ratio of total reads between genes
}
