% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0_RCA_readData_summary.R
\name{ISS_rotate}
\alias{ISS_rotate}
\title{Rotate and flip ISS data}
\usage{
ISS_rotate(data, rot = 0, flipby = NULL)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{rot}{Rotation angle (Degree). Default is zero. Negative angle define the anti-clock wise rotation.}

\item{flipby}{Flip the data points by x or y axis. Default is NULL. Possible value is NULL, "x", "y".}
}
\description{
Rotate by angle and flip by axix of ISS data
}
\examples{
mydata  <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                   cellid = "CellId", centX = "centroid_x", centY = "centroid_y")
mydata1 <- ISS_rotate (data = mydata, rot = 0, flipby = "y")
res     <- RCA_map(data = mydata, what = "cell")
res1    <- RCA_map(data = mydata1, what = "cell")

}
