% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2_ISS_rotate.R
\name{ISS_rotate}
\alias{ISS_rotate}
\title{Rotate and flip ISS data}
\usage{
ISS_rotate(data, rot = 0, flipby = NULL)
}
\arguments{
\item{data}{Input data in class MolDiaISS. Output of \link[MolDia]{readISS}.}

\item{rot}{Rotation angle (Degree). Default is zero. Negative angle define the anti-clock wise rotation.}

\item{flipby}{Flip the data points by x or y axis. Default is NULL. Possible value is NULL, "x", "y".}
}
\description{
Rotate by angle and flip by axix of ISS data
}
\examples{
# Read ISS data
mydata  <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                   cellid = "CellId", centX = "centroid_x", centY = "centroid_y")

# Flip/rotate ISS data
mydata1 <- ISS_rotate (data = mydata, rot = 0, flipby = "y")

# Plot ISS data
res     <- ISS_map(data = mydata, what = "cell")
res1    <- ISS_map(data = mydata1, what = "cell")

}
