% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCA_ClusterSelect.R
\name{RCA_ClusterSelect}
\alias{RCA_ClusterSelect}
\title{Select cluster of interest after clustering.}
\usage{
RCA_ClusterSelect(data, cluster_id = NULL)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readISS}.}

\item{cluster_id}{Cluster to select}
}
\description{
Select cluster of interest after clustering
}
