% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1.1_RCA_GridSelect.R
\name{RCA_GridSelect}
\alias{RCA_GridSelect}
\title{Select grid of interest from a tissue.}
\usage{
RCA_GridSelect(data, grid_id)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{grid_id}{Grid to select}
}
\description{
Select grid of interest from a tissue
}
\examples{
ex_data <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),cellid = "CellId", centX = "centroid_x", centY = "centroid_y", rpc = 3)
ex_data <- readRCA(file = "D:\\\\Project Sweden\\\\SU Project\\\\Working code\\\\Data\\\\All tissue\\\\CellBlobs2_1.csv",
                   cellid = "CellID", centX = "centroidX", centY = "centroidY", rpc =1)


}
