% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0_RCA_plot.R
\name{RCA_map}
\alias{RCA_map}
\title{Plot function for RCA_class data.}
\usage{
RCA_map(data, what = "cell", xlab = "centroid_x", ylab = "centroid_y",
  main = "Main plot", ptsize = 1, image = TRUE, live = FALSE,
  label.topgene = NULL, gene = NULL, cluster_id = NULL,
  same.y.lims = FALSE)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readRCA}.}

\item{what}{What to plot. Values can be "cell", "cluster", "tsne". Default is "cell".}

\item{xlab}{Label of x-axis}

\item{ylab}{Label of y-axis}

\item{main}{Main title}

\item{ptsize}{Point size}

\item{image}{Plot image. Default is TRUE.}

\item{live}{Plot interactive image. Default is FALSE.}

\item{label.topgene}{Active only when "what = cluster or tsne". Number of genes to label each cluster. Only work when data is clustered and clusted marker
has identified in cluster.marker slot of input data.}

\item{gene}{Gene of interest, Now limited to max 20 genes.}

\item{cluster_id}{Which cluster to plot. Only work when what = "cluster".}

\item{same.y.lims}{Set all the y-axis limits to the same values.}
}
\description{
Map RCA data based on cell, cluster or tSNE
}
\examples{
## Reading data
left_hypo <- readRCA(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"), 
                        cellid = "CellId", centX = "centroid_x", centY = "centroid_y")
left_hypo <- RCA_preprocess(data = left_hypo, normalization.method = "LogNormalize", 
                        do.scale = F, do.center = F)
left_hypo <- RCA_cluster (data = left_hypo, method = "seurat",resolution = 0.05)

## Plot violin plot
res1       <- RCA_map(data = left_hypo, what = "vlnplot", gene = left_hypo@gene[1:4], same.y.lims = F)

## Dimention reduction by tSNE on clustered data
left_hypo   <- RCA_tsne(data = left_hypo, do.label = TRUE, perplexity= 30)

## Plot tSNE cluster
res2       <- RCA_map(data = left_hypo, what = "tsne")

}
