% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCA_tsne.R
\name{RCA_tsne}
\alias{RCA_tsne}
\title{Dimentionality reduction to 2D by tSNE}
\usage{
RCA_tsne(data, clus = NULL, pc = NULL, perplexity = 100)
}
\arguments{
\item{data}{Input data in class RCA_class. Output of \link[MolDia]{readISS}.}

\item{clus}{As factor. Each cell names has id. Cell order of data and cell order of clus should be same. (work for class data.frame)}

\item{pc}{Desired percent of variance to be explained by PCA. Default in NULL , equavalant with 0.9.}

\item{perplexity}{Numeric; Perplexity parameter. See  \link[Rtsne]{Rtsne}

## param do.label Label tsne plot or not}
}
\value{
2D dataframe of points.
}
\description{
Any data in class RCA_class clusteded or not clustered used to reduce dimention to 2D by RCA-tsne.
}
\examples{
## Reading data
data_3 <- readISS(file = system.file("extdata", "Hypocampus_left.csv", package="MolDia"),
                  cellid = "CellId", centX = "centroid_x", centY = "centroid_y")
## Arrange marker gene
data(marker_gene)
marker_gene <- marker_gene
mark_gene <- list(genr = marker_gene$genr, neuron = c(marker_gene$genr_neuro,
                                                      marker_gene$genr_neuro_pyra1,
                                                      marker_gene$genr_neuro_pyra2,
                                                      marker_gene$genr_neuro_inter1,
                                                      marker_gene$genr_neuro_inter2,
                                                      marker_gene$genr_neuro_inter3,
                                                      marker_gene$genr_neuro_inter4,
                                                      marker_gene$genr_neuro_inter5,
                                                      marker_gene$genr_neuro_inter6),
                                            nonneuron = marker_gene$genr_nonneuro)

## Barplot of Neuronal marker gene and extract those cells only
neuron_group <- RCA_barplot(data = data_3, gene = mark_gene, gene.target = 2,
                            at.least.gene = 2, gene.show = 2)
## Data preprocessing
neuron_group <- RCA_preprocess(data = neuron_group, normalization.method = "LogNormalize",
                               do.scale = TRUE, do.center = TRUE)

## Cluster data based on SEURAT pipeline
neuron_group_clust  <- RCA_cluster(data = neuron_group, pc = 0.9, resolution = 0.3)

## Dimention reduction by tSNE on non-clustered data
tsne_noclust <- RCA_tsne(data = neuron_group)

## Dimention reduction by tSNE on clustered data
tsne_clust   <- RCA_tsne(data = neuron_group_clust, pc= 0.9, perplexity= 100)

## Get cluster marker
tsne_clust <- RCA_marker(data = tsne_clust, topgene =15, marker.sig = 0.005,
                                        test.use="bimod", main = "")
## Plot tSNE data
result <- RCA_map(data = tsne_clust, what = "tsne")

}
